Export routes to separate container, export redux configure store to separate file
